requirements.txt

applaunchservices==0.2.1
argh==0.26.2
asgiref==3.3.1
astropy==4.0.1.post1
atomicwrites==1.4.0
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.1
bkcharts==0.2
bleach==3.1.5
boto==2.49.0
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
contextlib2==0.6.0.post1
cryptography==2.9.2
cycler==0.10.0
cytoolz==0.10.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.3
django-debug-toolbar==3.1.1
django-phonenumber-field==5.0.0
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
fsspec==0.7.4
future==0.18.2
glob2==0.7
gmpy2==2.0.8
greenlet==0.4.16
h5py==2.10.0
HeapDict==1.0.1
imagesize==1.2.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
json5==0.9.5
jupyter==1.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.5
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
mock==4.0.2
more-itertools==8.4.0

nbformat==5.0.7

numba==0.50.1
numexpr==2.7.1
olefile==0.46
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
partd==1.1.0
path==13.1.0
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
phonenumbers==8.12.13
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
ptyprocess==0.6.0

pycodestyle==2.6.0
pycosat==0.6.3

pycurl==7.43.0.5

pyflakes==2.2.0


pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
QtPy==1.9.0
rope==0.17.0
seaborn==0.10.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1

tables==3.6.1
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4

toolz==0.10.0
tornado==6.0.4

traitlets==4.3.3
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.9
xlwings==0.19.5
xlwt==1.3.0
xmltodict==0.12.0
zict==2.0.0
zipp==3.1.0
zope.event==4.4
